requirements.txt
python-dotenv
azure-ai-inference
openai